requirements.txt
streamlit>=1.33.0
pillow>=10.3.0
litellm>=1.41.10
python-dotenv>=1.0.1
requests>=2.31.0
ollama>=0.3.0
